Move json and xml parsers to core
