+++ libs/langchain/langchain/output_parsers/format_instructions.py
@@ -47,21 +47,6 @@ Please follow the standard YAML formatting conventions with an indent of 2 space
 
 Make sure to always enclose the YAML output in triple backticks (```)"""
 
-XML_FORMAT_INSTRUCTIONS = """The output should be formatted as a XML file.
-1. Output should conform to the tags below. 
-2. If tags are not given, make them on your own.
-3. Remember to always open and close all the tags.
-
-As an example, for the tags ["foo", "bar", "baz"]:
-1. String "<foo>\n   <bar>\n      <baz></baz>\n   </bar>\n</foo>" is a well-formatted instance of the schema. 
-2. String "<foo>\n   <bar>\n   </foo>" is a badly-formatted instance.
-3. String "<foo>\n   <tag>\n   </tag>\n</foo>" is a badly-formatted instance.
-
-Here are the output tags:
-```
-{tags}
-```"""
-
 
 PANDAS_DATAFRAME_FORMAT_INSTRUCTIONS = """The output should be formatted as a string as the operation, followed by a colon, followed by the column or row to be queried on, followed by optional array parameters.
 1. The column names are limited to the possible columns below.
